requirements.txt
Flask==3.0.0
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
feedparser==6.0.10
Werkzeug==3.0.1
reportlab==4.0.4
gunicorn==21.2.0
